dumping work so far
