+++ laff/functions.py
